Completely revamped core module dependency handling.
We now create empty packages for core modules to simplify
dependency handling down the line, and do some version checks to see
if we need to create packages for newer versions of core modules
required by other modules.

The basic idea now is that each core module recipe is only good for
a particular version of Perl. Whenever a new version of Perl comes out,
we're going to need to update the recipes for this new version. I guess,
really this should be handling using environment variables and build
strings like CONDA_PY does, but I haven't had a chance to do this yet.