Make conda convert automatically create platform subdirectories
It now writes the packages to platform subdirectories, like
win-32/newpackage.tar.bz2. This allows to specify more than one platform in
the same command, which was previously impossible because the packages all
have the same name and only one output directory was given. -o is no longer
required and defaults to .