Add conda_preserve_egg_dir
I wasn't sure if this should just use zip_safe, but I chose to separate it,
because they seem to be doing separate things. Also, setuptools wants to use
zips when it can (although it is conservative in its default checks), and
conda does not, so it probably makes sense to keep them different if just
because of the different defaults.