Don't use "metadata.name" as an import to test with bdist_conda
My understanding is that "packages" is already a required argument in
a setup.py to set packages to install, so I don't think there is any
benefit to adding "metadata.name" as a test import. If the name of the
project is a valid import, it should already be included in "packages".