Fake conda_build.config as an object
This lets us mutate properties on the object, and have them propagate to other
properties (because we can use @property on methods of it).  For instance, we
can change the build prefix depending on aspects of the meta.yaml, and have
other paths that are relative to the build prefix be updated accordingly.

This maintains backwards compatibility with all syntax for importing things
from conda_build.config, but it is best to always just import the module, like

import conda_build.config

and access things as attributes, like conda_build.config.CONDA_PY, rather than

from conda_build.config import CONDA_PY

as the latter will not be updated if somewhere else in the code mutates
conda_build.config.CONDA_PY.