Handle symlinks in built packages
Symlinks to paths outside conda-bld are fine. Symlinks to paths in the build
prefix are fine, but they must be relative. The post script now makes these
symlinks relative if they are absolute. Symlinks to paths in conda-bld outside
the build prefix (such as the source directory) are disallowed and give an
error.

Fixes #97.