Introduce build/use_destdir option in meta.yaml
This introduces a build/use_destdir option to meta.yaml.  When this
option is set to true files should be installed into $DESTDIR
instead of $PREFIX.  Inside $DESTDIR the $PREFIX should still be
respected, e.g. $DESTDIR/$PREFIX/bin/my_binary is a valid file.

This is useful in case one needs to build a package which
build-depends on itself.  In this case the normal build system would
not detect the new files added.