Revert "Fix PATH in test environment in Windows"
This reverts commit 46b5f971174d03b580057ae874924d03a84b5f52.

This issue was in response to #41, but this is *NOT* what we want. vcvarsall
adds things to the PATH, and without PATH as a variable, we completely clear
those changes.

We could also fix this by moving the vcvarsall call below the set commands in
the bld.bat, but this was is preferable because it keeps the build prefix at
the very front of the PATH.