Use find_lib() in osx_ch_link() (first pass)
This should make libraries point to the correct locations for their dependent
links, even if they are in completely different locations. Currently, the
logic is only correct if the dependent library sits alongside the library that
links against it.

All the logic of when osx_ch_link() should not do anything is now delegated to
find_lib().

Also removed special casing of /usr/local/lib/libgcc_s.1.dylib, since we now
have a GCC package in the Continuum repos, so if someone is linking against a
different GCC, it should be assumed that they know what they are doing.