Fix regression in always_include_files introduced in 1.12.0
Starting in version 1.12.0, this switched from including files if
present to requiring those files to be present.  This breaks the
conda-env build and any other build that used the original design.

Given that it is a backwards in-compatible change, this a break without
bumping to version 2.0.0.  Reverting this change and changing to a
simple error message.

Original Commit: f708dd0225b5292b644133afd190800e24e5e047