skeleton pypi: add --manual-url argument
Currently, when conda-skeleton has more than one sources to choose from, it asks
the users for the decision. The patch changes this behaviour to download from
the source with the least source size. This is because this former choice can
only add to the confusion of a _normal_ user. An informed user can override the
behaviour with the newly added `--manual-url` command line argument.

Signed-off-by: Harsh Gupta <gupta.harsh96@gmail.com>