Don't rename meta.yaml to meta.yaml.orig
The parsed meta.yaml information is already in recipe.json. The yaml dumped
meta.yaml confuses the conda-build selector parser, so that the recipe cannot
be rebuilt as is (and anyway, the meta.yaml in info/recipe should be the same
meta.yaml that was used in the original recipe).