fix import_tests: module name != package name
conda-skeleton implicitly assumes that package name is same module name and
adds an import test for that, it is not necessarily the case. Examples:
the package dnspython installs a module named dns and the package factory_boy
installs the package factory. The earier behavior was causing build failures
for such packages.

Signed-off-by: Harsh Gupta <gupta.harsh96@gmail.com>