Automatically remove a package of the recipe itself if it is installed as a build dependency
Otherwise, if the version installed as a build dependency has the same files
as the version being build, a package will be built that don't include those
files.

One can attempt to work around this in the build script by installing some
build dependencies manually, but this is messy will result in a package that
includes its dependencies if not done correctly.