Make rpath added by conda develop an absolute path
During conda build process, the package is placed inside site-packages
of the environment so it makes sense for @rpath to be a relative path.

However, conda develop adds the source path to conda.pth which could
be anywhere with respect to conda environment. conda develop also
relinks shared objects against the libraries installed in the
environment. There is no reason to make @rpath relative to source
folder since source folder could be anywhere, but making it an absolute
path does make it more readbale and makes the source folder
relocatable.