Add git_recursive option (defaults to true)
Since the change in #745, the requirement to use git_url instead
of path even for local git repositories forces a clone of any
submodules every time a conda build is run (the --mirror option
used to create the local cache does not mirror submodules). This
can be a problem if the submodules are large (e.g. several
hundreds of MB is not uncommon) and especially if they are not
actually required for the build.

This commit adds a git_recursive option, which defaults to true.
When true, the git clone that conda build makes is recursive, and
submodules are included; when false, the clone is not recursive
and submodules are not cloned.