Stop setting values for undefined variables listed in `script_env`.
In a recipe, a user can specify variables that can be defined externally
to influence the recipe. The previous practice was to leave these
undefined. However, a commit was added that began setting them to
`<UNDEFINED>`. If a user intentionally left these variables undefined,
they must now go back and unset them afterwards. The intent of this was
likely to smoke out errors. This could initially be easily worked around
in bash or batch. However, 2 pass jinja templating has forced us to
handle this inside the `meta.yaml`, which is more challenging and
unnecessarily ugly. Instead of continuing this trend of setting
intentionally undefined variables, this breaks from that change and
simply warns the user when they have an undefined environment variable.