Allow default values for environment variables only if specified.
Instead of simply leaving undefined environment values as such. Provide
a mechanism for giving them a default value in `meta.yaml`. This makes
the breaking nature of this change less severe. Users requiring or
desiring to set the environment variables to some default value when not
picked up externally is possible. However, this is better than the
original strategy as this does it in a manner that is explicit.