Add new option --keep-old-work
The point is to allow engineers to debug issues with callstacks spanning
multiple conda packages without having to hack conda-build or mess around
with multiple miniconda environments.

The contents of the old work directory are moved into a temporary dir
before the build commences and moved back after it has completed. It was
done this way because the source unpacking logic is fairly hostile to
this kind of change.