Attempting to improvng the logic for VS on Win
Both VS9 and VS10 do not work out of the box for 64-bit builds.
This should try and remedy that. For VS9 it prefers the
Microsoft Visual C++ Compiler for Python 2.7 and when
falling back VS9 will call the correct bat file manually, rather
than relying on the broken logic in vcvarsall.bat.

For VS10, it will correctly call the Windows SDK rather than
relying on the also potentially incorrect vcvarsall.bat.