test_build_recipes.py: customize recipe builds by adding _driver.sh
test_build_recipes.py will check if _driver.sh exists in the recipe
directory before invoking the build command (i.e., a variant of
`conda build .`). If _driver.sh exists, it will instead invoke it, passing
it the build command as arguments. This allows the recipe enviromnent to be
customized; for example:

  #!/bin/bash -e
  ... autogenerate some files the recipe may depend on ...
  "$@"
  ... perhaps do some clean up ...