Refactor out paths in windows for accurate tests
Instead of having the paths defined twice, just have them defined
once at the top of the file so that they can be imported
from the tests. This reduces the probability of error and since
the paths are defined at runtime it doesnt matter if they are
created when the windows.py file is parsed.