Include the run specs in the test environment
Due to what looks like a simple typographical or historical error, the command that creates the test environment is not explicitly requesting the packages listed in the `requirements/run` of the `meta.yaml` file in the environment creation. 

Of course, because they are included in the newly-built package's metadata, you would think this doesn't matter---they're going to get picked up due to that package's dependencies.

The problem is that conda priorities maximizing the versions of the packages that are explicitly requested *first*, and then maximizes the versions of any packages included by dependency *second*. During the *build* stage, then, the build dependencies are given high version-maximation priority; and during the *test* stage, they were given lower version-maximization priority.

This sometimes caused `conda` to select different version of the packages between `build` and `test` environments. Technically, if this causes functional differences, those differences should be investigated, and tighter run specs should be considered. Still, it causes confusion and should be fixed.