Improve Windows MSVC logic
Use setuptools/distutils to do the heavy lifting. Instead of needing all
the logic around the Microsoft C++ Compiler for Python 2.7, we just use
setuptools which already supports this. We do however, maintain the shim
for x64 builds if VS2008 is required. This should also fix up the Python
3.3/3.4 builds since there is better logic for detecting if the Windows
SDK failed (checking an environment variable).