source: Allow 'relatively pathed' git submodules
Git allows submodules to be added with either a relative or an
absolute URL.

Absolute URLs work fine as-is, but relative ones do not. This
is because we mirror the git repositories and, after that the
relative URL is taken as relative to where our mirror lives.

To work around this, any submodules specified this way need to
be cloned to the same location relative to the parent module.

This must be done recursively.

For an example of a git repository that fails to clone without
this change, see: https://code.qt.io/qt-creator/qt-creator.git
Here, `qbs` is a relatively specified submodule.