Make `CPU_COUNT` an automatic pass through environment variable.
In some situations, users find the processor information to be reported
incorrectly by the infrastructure they may be running a build on. For
instance, a user may be running a build on a node where the node
supports more cores than the user is actually provided. As a result, the
detection of the number of cores will be too high and cause issues
completing the build.

Previously it was thought `MAKEFLAGS` would be an acceptable way to
accomplish this. In fact, that strategy does work. Most of the cases
where a parallel build is desired are using `make` and thus can easily
leverage this. Though it doesn't really catch all cases and so something
else still needs to be done for them. Also, using it in this way can
result in an implicit setting of parallel builds. Thus there were some
cases where a parallel build was not reliable. As a result, users had to
unset `MAKEFLAGS` entirely. This is problematic as other flags not
related to parallelism may be lost in the process.

The strategy proposed here is to allow the environment variable
`CPU_COUNT` to be automatically passed through to the recipe. This
allows users to opt-in to parallelism if it helps their build. However,
if their build does not work well in parallel it is not forced upon
them. As it simply overrides an existing environment variable used to
specify the number of cores available, recipes that use this feature can
be safely built in different environments by setting this environment
variable to the appropriate value in that setting. Further strategies
build strategies that don't use `make`, but do have the ability to build
in parallel can be set appropriately using the `CPU_COUNT` environment
variable. Finally, by allowing `CPU_COUNT` to be passed through we
provide a useful tool in debugging recipes. Instead of having to tinker
with hard-coded values in recipes, a user can set this environment
variable to different values externally and see how it affects there
build.

In short, this proposal should hopefully provide a useful feature for
those eager to support parallel builds by putting that control more
firmly in the hand of the users. Further, this should help separate what
is appropriate for build infrastructure from the recipes themselves
resulting in greater flexibility.