source.py: Use the build env for all `subprocess` calls
Python defaults to using the initial/calling environ in
the `subprocess` module for some reason I can't fathom.

Instead of adding `env=os.environ` to each function two
new functions are provided: `check_call_env()` and also
`check_output_env()`. If `env` is passed to these, they
simply pass the value on, otherwise `env` gets added to
`kwargs` with the value of `os.environ`.

With this change and the previous one, it is possible to
build (some?) packages that use `git` without needing it
to be pre-installed in any system location. For example,
on Windows, adding `m2-git` and `m2-filesystem` as build
requirements is sufficient.

This also fixes up test_checkout_tool_as_dependency which
was not putting the bad svn executable on the PATH making
the test that the build dependency svn got added invalid.