Add build/skip_compile_pyc option
Some packages ship .py files that should not be compiled because either they
cannot be compiled (for example they contain templates) or should not be
compiled yet because the Python interpreter that will be used cannot be known
at build-time.

An example of such a package is Qt Creator which uses GDB's embedded Python
interpreter to give a richer debugging experience. This package falls foul of
both scenarios given above.