retry clone with local mirror as reference if first from-mirror clone fails
While from-local clone will fail if the repo contains files in LFS,
the local mirror can still be used as a reference to avoid redownloading git objects.

The `--reference` clone is only used if the normal mirror clone fails
because it's slower, and will hopefully become obsolete with future LFS versions.