Copy CONDA_PATH_BACKUP into build/test env.
Since https://github.com/conda/conda/pull/3175 is in flux* we must
detect whether this environment variable is being used by the conda
coda-base on this platform.  Since this is on the interface between
conda and conda-build this check is performed in conda_interface.py.

* one of the nice things about that PR is that, AFAIK, it makes the
behaviour in-line with Windows since https://github.com/conda/conda/pull/2917