Add two new entry points to convert test case
One of the new entry points is the same as the previous, only with
the last characters removed, and the other is is the same, but with
additional characters.

This will in most cases trigger a bug where substrings are replaced
in longer paths. For this to happen, a file with path equal to another
except for some extra characters at the end must end up later in the
file the tar file. Since the order of the members seem to be random
(at least implementation dependent) the bug is not guaranteed to be
triggered.

With this commit, it is sufficient that one of the files
    bin/test-script-setup.py
    bin/test-script-setup.pyc
    bin/test-script-manual
    bin/test-script-manual-postfix
end up later in the tar file than
    bin/test-script
for the bug to trigger.