conda_build/windows.py: avoid trailing semicolon in %MSYS2_ARG_CONV_EXCL%
The previous method of setting the %MSYS2_ARG_CONV_EXCL% environment variable
tried to preserve any preexisting value of the variable. Not only does this
make the build process depend on the initial environment more, it generally
leads to a trailing semicolon in the variable value. This seems to cause MSYS2
to parse the variable as having an empty value that prevents *all* argument
conversions, as investigated in the conda-forge `cairo` package:
https://github.com/conda-forge/cairo-feedstock/pull/18.

I do not know what fallout this change will have for existing build scripts,
but the current behavior cannot be what was intended.