metadata: Do not set exclude_from_build_hash
We're not sure how to proceed with this feature. We no longer need to use it
for the 'drop dependencies' subpackage thing since we are not doing it that
way anymore (instead you can declare requirements for subpackages.

It could go back to its original purpose of loosening mkl and numpy deps when
the C ABI is not used directly (can min_pin and max_pin not do that though?).

Here it was preventing the just-before-build finalize_metadata call from finding
pin_subpackages in other_outputs because the config.variant used as part of their
key didn't contain this while the metadata.config.variant did contain it.