jinja: Rename pin_subpackage(name_only => stringify_subpackage_pins)
The jinja2 expression becomes a string so that we can (if necessary) decode it later.
Mike thinks this may not be necessary but I do not see an easy way of getting this
information again if it gets turned into just the subpackage name; I guess you could
look, in an un-rendered version of the metadata for a dependency with that name to
get the values back, but this seems more direct.

The following commit will pull apart this string and call the jinja2 template function
to swap this jinja2 macro encoded as a string back into a normal spec string (with
exact pinning, if that's what the recipe asked for).

There's a good chance I'm not fully understanding a part here, but doing it this way
would work anyway even if it's carrying more info than we need, and is IMHO, easier
to debug and (for me to) reason about.