render: Swap the jinja2 macro (pin_subpackages) encoded as a string back into a normal spec string
We use pin_dependencies() to do this which, at present, gives us the wrong hash because it
looks this up from other_outputs which has not yet been finalized.

Mike said he would look into finalizing it around the time it gets created. For this to work,
the source download needs to be done before finalizing.

Alternatively, we can keep track of what subpackages we have built so far and do the pinning
against that information.