Relax entry_points syntax
Allows elements of the build/entry_points section of a recipe to contain Python
entry points without spaces between the name of the console script and the
module path and function.  For example, foo=foo_module.func.

This syntax was previously allowed for some recipe but not noarch: python
recipe.

closes #1893