Fix mmap/file problems on VirtualBox shared folders
In one case, we can open the file as read-only before mmapping it to
prevent OSError 22.

When that fallback fails, set mm to the file contents not the file
handle when mmap fails to prevent:

    m = SHEBANG_PAT.match(mm)
TypeError: expected string or bytes-like object

In the other case, the file is written to via the memory mapping so
there use flags=mmap.MAP_PRIVATE instead.