'target' packages must not set 'host_subdir'
.. as otherwise the that package's dependencies are looked for in 'target'
 subdir which we never want.

'target' packages (e.g. libstdcxx, libgfortran) cross the rubicon in that
they are to be run on a machine different from the build/host so they can
not depend on packages for that same machine.

Setting the subdir for these needs to be done as a very late-stage thing
that the solver never gets to hear about.