Add new functionality to --keep-old-work flag
The --keep-old-work was deprecated and ended up having the same functionality
as the --dirty flag. However, there are times when a developer would want to
have the work directory remain intact even after a successful build/test.
Rather than create a new flag, the --keep-old-work flag will be used for
this particular case.