skeleton(rpm): Various bits and pieces (WIP)
1. Trim whitespace and newlines in description and summar

Leading whitespace can cause such lovely errors as:

```
Unable to parse meta.yaml file

Error Message:
--> while parsing a block mapping
-->  in "<unicode string>", line 16, column 7
--> did not find expected key
-->  in "<unicode string>", line 22, column 9
```

2. skeleton(rpm): Add a way to specify missing dependencies (WIP)

Passing the package name as package,missing_dep will cause missing_dep
to be added as a dependency of package.

This is needed because distros sometimes forget to list always-present
dependencies, though in fairness, there could be something missing in
my repository parsing since I reverse engineered it.

Though in the case that caused this to be added it does look to be just
missing. That case is libxcb-devel,libxcb1 in Aarch64 openSUSE Leap 42.3

This is not a good way to do this, it has a race condition where a
libxcb-devel recipe added as a dependency of another package would not
get this extra dependency added. A better idea is to add:

```
--dependency-add=libxcb-devel,libxcb1
```

.. so that any time libxcb-devel is seen, the dependency gets added.

3. Set as `noarch: generic` instead of `target: noarch`

4. Use the version and epoch of the provider of the dependency.