cran-skeleton: allow certain blank keys
The 'Writing R Extensions' manual doesn't mandate that if there are no
suggestions for a package, then the field should not be present. It can
still be present, with a blank value, for example:
https://cran.r-project.org/web/packages/corpcor/DESCRIPTION

This patch adds the functionality to not throw an error during parsing
if the value of certain fields is empty. It should just ignore that
field and continue.