implement subspace selection
With zip_keys, it was very frustrating to have a central config file that
specified several zip_keys (python, vc, c_compiler, etc.), but to not allow
easy reduction of that build matrix by other config files.  One variable would
get trimmed to the shorter set of values, but then conda-build would fail with
errors about mismatching zip_keys dimensions.

This commit filters out mismatching entries, but does so in a way that allows
the dimensionality of the composite configuration to be maintained.  This also
allows things like

``conda build somerecipe --python=2.7`` to work when larger python build
matrices are defined in conda_build_config.yaml