Add a `recipe_url` to `about`
If Anaconda supports this some day, it will make it easier for people to
navigate from the package back to the recipe source. Seems pretty handy
if one finds the package has an issue and would like to figure out why.
Could also be helpful for someone that is curious.