Added "Error:" to several exception messages
Motivation: It was unclear that conda build failed to me (why), because it was not obvious that "Error: Received dictionary as spec.  Note that pip requirements are not supported in conda-build meta.yaml." is an error message.
Fix: added "Error: " string to several error msgs here which were not stating that.