Fix script renaming when converting from win to unix
When renaming files that end with -script.py, the update_executable_path
function looks for files that end with -script.py and should remove
the -script.py suffix from the file. Currently, only the .py suffix
is being removed. This commit fixes this issue by removing the entire
-script.py suffix.

Fix how the script file prefix is renamed