Don't rename executable paths in paths.json
When checking an executable to see if it should be renamed,
the checks consist of binary file checking and hidden file
checking. If the executable is a binary or a hidden file,
it is not renamed. While not renamed, the executable's
path is changed in paths.json when it shouldn't be. This
commit adds these checks when adding the executable path
to paths.json.

Move executable check from update_paths_file to update_executable_path

Remove redundant checks in update_executable_paths

Fix indentation in update_paths_file