no do disable pip's cache directory by default
Do not disable pip's cache directory.  When the cache is disable pip does not
try to build wheels for an install.  This is undesirable as it leads to
.egg-info directory where as wheel produce modern .dist-info directories.

This change allows the user to choose if pip's cache directory should be
enabled or not via the build script.  Pip's default is to enable the cache.