Add support for generating .tar.zst as well as .tar.bz2
To enable this CONDA_TARBALL_EXTENSIONS needs to be changed in conda,
can we make this value configurable via an env. var for now @kalefranz?
It would aid in testing this stuff.

We also remove utils.unzip here since we have tar_xf instead. It seems
unzip was basically unused anyway.