Add new `source/path_via_symlinks` metadata option
When used in conjunction with `source/path` (it does not make sense
otherwise) and with `source/folder` (trying to use it without that is
fatal for reasons of not wanting to destroy your source code) you can
keep your source-code where ever you want.

source/folder (trying to use it without that is fatal) you can
and with source/path (it doesn't make sense in any other mode), you can use this to have a
very pleasent workflow.