Add TEST_FAILED feature
Sometimes (esp. on Windows) scripts can fail without returning a errorlevel
(if a batch file exits before it has run any programs (builtins do not count
at least in most cases)).

We had an issue where some tests in the conda recipe were not being run because
we were not CALLing conda, and it would exit and we'd exit the whole test script.

This provides a way, on a case-by-case basis, to prevent this from happening. The
design is simple. At the start of your tests, you touch TESTS_FAILED, and at the
end of them you remove this file. If the tests exit before this happens we detect
the file and call it a failure.