rework extract cache to make reuse less direct
There was a problem where .tar.bz2 and .conda files were having their size, md5 and sha256 values get mixed up with their counterpart.  I'm pretty sure this was because of the shared cache thing.  We do need separate files (for the sake of different size/md5/sha256 values to be loaded), but we need not extract it from the original file.  We can copy it from the faster one (.conda) and then add in the correct size/md5/sha256 for the .tar.bz2, with the appropriate, non-overlapping filename.