Drop source RPMs (as they are unused currently)
Also community members have requested having checksums for the source
RPMs as well. While there is some rough idea as to how this might be
done, it appears to be a fair bit of work that no one has volunteered to
take on. To streamline usage a bit, the suggestion here is to leave
source RPMs out of the skeleton until we have solved this checksum
issue.