Revert to using corelist and LooseVersion
corelist is the cannonical method for determining whether
a perl module is included in the perl core library. The
metacpan community does not have a method for their API
to convey whether a module is in the core or not (
see metacpan/metacpan-api#925).  Until metacpan can
tell us whether a module is in the core, revert back to
using the corelist script that ships with perl.

Also revert to using distutils.version.LooseVersion for
version comparisons.  LooseVersion implements versioning
that is free-form like Perl and the pkg_resources.parse_version
will incorrectly sort versions compared that align with
PEP-440 against those that don't.  (Reading the code
pkg_resources assumes that all versions that do not comply
with PEP-440 occur before those that do. Since nothing in
the Perl world will intentionally comply with PEP-440
using pkg_resources will lead to confusing results.

When moving back to LooseVersion, we must fix issues like
the one reported in #994:
  TypeError: expected string or buffer
and it's Python3 equivalent that uses bytes instead of buffer.
This TypeError is encountered when we pass an int into
LooseVersion instead of a string or bytes.  Metacpan has
an open issue (metacpan/metacpan-api#789) that reports
inconsistancy in the returned type of versions in the json.

To address this, we now wrap distutils.version.LooseVersion
in our own LooseVersion helper that coerces the supplied
version parameter into a string.  We also strip any leading
'v' to avoid TypeError about comparing int against str.