Clarify 'deletes the build env' in concepts/recipe
I don't see anywhere in the docs where it is explained why the SOURCE_DIR, and build/host environments are deleted at the beginning of test phase (or in between build and test phase).  This commit captures that reasoning so that I can point here for new people learning about conda-build.