Respect keep_old_work, do not rm_rf(self.build_folder) when either it or dirty are set
.. same thing with host prefix.

The code and concepts behind --dirty, --keep-old-work and also
perhaps build/no_move_top_level_workdir_loops seem a bit muddled and
overlapping. --dirty should *only* deal with finding a new croot, and
have no part in whether to keep stuff around or not. That should be
left to a new option, like:
--keep=work,build-prefix,host-prefix,test-prefix,test-tmp
where --keep-old-work is deprecated but enables all of these.
Also, 'keep' might not be the best word, call it what we do,
--retain-via-rename=work,..