Fix test_copy_read_only_file_with_xattr, it fails on modern Linux
Problem is tmpfs doesn't tend to support xattrs. Instead we add a new fixture, testing_homedir which uses the homedir
instead. Also we throw a try: except: around the call to setfattr in this location and consider that an xfail.

This means this functionality will continue to be tested on more modern systems and will not flag false positives.

If for some reason we cannot create a test folder in expanduser('~') then that is also an xfail.