Many improvements to overlinking detection
Changes ::

Uses lief 0.10.1 now (no code changes were needed for this update)
Fixes _map_file_to_package for finding '.dylib' and '.so' without suffixes beyond that
Added our FakeDist to run-prefix, improve library_nature, do not fail if ccache exists
Fixes detection of where packages comes from.
  We must avoid using 'canonical' channel names (such as 'default', needing
  instead 'pkgs/main') when determining which channels packages come from.
  To do this, we use a new function, linked_data_no_multichannels().
  Let's hope we never end up comparing these with ones gotten via linked_data()
Add local_output_folder to get_build_index()'s locally cached things.
Ensure we pass 'channel_urls' as a list and add the 'local' channel to
  that.
Disambiguate 'prefix' DSO files in the build/host throughout ovelink detection.
Change ERROR :: {} not in prefix_owners to a more helpful warning and add a comment
  though it might be better to just remove this warning and let the final info
  detail any issues (for they will show up as overlinking errors).

Notes ::

This is still not the speed boost that this stuff needs, so I am hoping the update to
0.10.1 helps with that. Looking at the history, it should, somewhat.

Case-insensitive lookup for Windows (hopefully only for files in the sysroot - which
is being referred to as the whitelist in the messages ATM unfortunately - still remains
to be done (and will probably get squashed into this commit).