Revert a part of my post/overlink fixes so it does not hold up the release
The downside is that some packages may not get picked up. I need to check that though. It cannot remain this broken.

The problem seems to be that depending on how the channel was identified to conda when it made the Dist is how it will
remain and one created from 'defaults' is not equal to one created from 'pkgs/main'. During conda-build post, our channels
seem to come from 'pkgs/main' but at other times they come from 'defaults'.